refactor: Updating database and table structures (#37)
* refactor: created migration file to update transaction table

* refactor: created migration file to update subscriptions table

* refactor: updated column order for certain tables

* refactor: created migration file to update accounts table

* refactor: updated account and authentication servies

* refactor: commands folder now follows app structure

+++ Expensier.Domain/Services/Subscriptions/SubscriptionService.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static Expensier.Domain.Models.Subscription;
 
 namespace Expensier.Domain.Services.Subscriptions
 {
@@ -12,40 +13,85 @@ namespace Expensier.Domain.Services.Subscriptions
         private readonly IDataService<Account> _accountService;
         private readonly IDataService<Subscription> _subscriptionService;
 
-        public SubscriptionService(IDataService<Account> accountService, IDataService<Subscription> subscriptionService)
+
+        public SubscriptionService( IDataService<Account> accountService, IDataService<Subscription> subscriptionService )
         {
             _accountService = accountService;
             _subscriptionService = subscriptionService;
         }
 
-        public async Task<Account> AddSubscription(Account currentAccount, string companyName, string subscriptionPlan, DateTime dueDate, double amount, SubscriptionCycle subscriptionCycle)
+
+        public async Task<Account> AddSubscription(
+            Account currentAccount,
+            string companyName,
+            string subscriptionPlan,
+            double amount,
+            SubscriptionFrequency subscriptionFrequency,
+            DateTime dueDate )
         {
             Subscription newSubscription = new Subscription()
             {
-                AccountHolder = currentAccount,
-                CompanyName = companyName,
-                SubscriptionPlan = subscriptionPlan,
-                DueDate = dueDate,
+                User = currentAccount,
+                Name = companyName,
+                Plan = subscriptionPlan,
+                Frequency = subscriptionFrequency.ToString(),
                 Amount = amount,
-                SubscriptionType = subscriptionCycle.ToString()
+                IsActive = true,
+                DueDate = dueDate,
             };
 
-            currentAccount.SubscriptionList.Add(newSubscription);
-
-            await _accountService.Update(currentAccount.ID, currentAccount);
+            currentAccount.SubscriptionList?.Add( newSubscription );
+            await _accountService.Update( currentAccount.ID, currentAccount );
 
             return currentAccount;
         }
 
-        public async Task<Account> DeleteSubscription(Account currentAccount, Guid subscriptionID)
+
+        public Subscription GetSubscriptionByID( Account currentAccount, Guid subscriptionID )
         {
-            currentAccount.SubscriptionList
-                .Remove(currentAccount.SubscriptionList
-                .FirstOrDefault((subscription) => subscription.ID == subscriptionID));
+            return currentAccount.SubscriptionList.FirstOrDefault( ( subscription ) => subscription.ID == subscriptionID );
+        }
 
-            await _accountService.Update(currentAccount.ID, currentAccount);
 
-            await _subscriptionService.Delete(subscriptionID);
+        public async Task<Account> DeleteSubscription( Account currentAccount, Guid subscriptionID )
+        {
+            Subscription subscriptionToDelete = GetSubscriptionByID( currentAccount, subscriptionID );
+
+            if ( subscriptionToDelete != null )
+            {
+                currentAccount.SubscriptionList?.Remove( subscriptionToDelete );
+
+                await _accountService.Update( currentAccount.ID, currentAccount );
+                await _subscriptionService.Delete( subscriptionID );
+            }
+
+            return currentAccount;
+        }
+
+
+        public async Task<Account> AcivateSubscription( Account currentAccount, Guid subscriptionID )
+        {
+            Subscription subscriptionToActivate = GetSubscriptionByID( currentAccount, subscriptionID );
+
+            if ( subscriptionToActivate != null )
+            {
+                subscriptionToActivate.IsActive = true;
+                await _accountService.Update( currentAccount.ID, currentAccount );
+            }
+
+            return currentAccount;
+        }
+
+
+        public async Task<Account> CancelSubscription( Account currentAccount, Guid subscriptionID )
+        {
+            Subscription subscriptionToCancel = GetSubscriptionByID( currentAccount, subscriptionID );
+
+            if ( subscriptionToCancel != null )
+            {
+                subscriptionToCancel.IsActive = false;
+                await _accountService.Update( currentAccount.ID, currentAccount );
+            }
 
             return currentAccount;
         }
